build can be continue

+++ rom/lede/make.sh
@@ -6,10 +6,14 @@ test -n "$IDXS" || IDXS="0"
 
 test -n "$CONFIG_VERSION_NUMBER" || CONFIG_VERSION_NUMBER="3.0.0_build`date +%Y%m%d%H%M`"
 
+test -f .build_ptpt52/env && source .build_ptpt52/env
+
 echo build starting
 echo "CFGS=[$CFGS]"
 echo "IDXS=[$IDXS]"
 echo "CONFIG_VERSION_NUMBER=$CONFIG_VERSION_NUMBER"
+mkdir -p .build_ptpt52
+echo "CONFIG_VERSION_NUMBER=\"$CONFIG_VERSION_NUMBER\"" >.build_ptpt52/env
 sleep 5
 
 find feeds/luci/ -type f | grep -v .git\* | while read file; do
@@ -29,7 +33,8 @@ for i in $IDXS; do
 	touch ./package/base-files/Makefile
 
 	for cfg in $CFGS; do
-	set -x
+		test -f .build_ptpt52/$cfg && continue
+		set -x
 		cp feeds/ptpt52/rom/lede/$cfg .config
 		sed -i "s/CONFIG_VERSION_NUMBER=\".*\"/CONFIG_VERSION_NUMBER=\"$CONFIG_VERSION_NUMBER\"/" ./.config
 		sed -i "s/CONFIG_VERSION_DIST=\".*\"/CONFIG_VERSION_DIST=\"$CONFIG_VERSION_DIST\"/" ./.config
@@ -39,8 +44,11 @@ for i in $IDXS; do
 		set +x
 		test -n "$1" || exit 255
 		$* || exit 255
+		touch .build_ptpt52/$cfg
 	done
 done
+
+rm -rf .build_ptpt52
 
 build_in=$(cd feeds/ptpt52/rom/lede/ && cat $CFGS | grep TARGET_DEVICE_.*=y | sed 's/CONFIG_//;s/=y//' | wc -l)
 build_out=$(find bin/targets/ | grep -- '\(-squashfs\|-factory\|-sysupgrade\)' | grep -v factory | grep ptpt52 | grep -v root | grep -v kernel | sort | wc -l)
